Changed ant.png to a better ant for future purposes. Added Constants.java to better keep constants for future development.
Added a world to MainScreen.java. This world is where our game will play out. So far, there is a box (world border, size adjustable via Constants.java) that contains 4 boxes (acting as up, down, left, right walls for the border).

There is also a box for the player. The player box is able to move around with WASD or arrow keys. The player is able to not pass through the border box.

The player also has a camera attached to always keep the player centred.

There are a variety of movement methods we can use. Checkout MainScreen.inputUpdate to modify CameraChoice to test different movement mechanics to see which one we should use. (I personally vote physics without momentum).

Also removed Image actor, not needed AFAIK.